% Auto-generated: do not edit by hand
\name{dashJsmole}

\alias{dashJsmole}

\title{DashJsmole component}

\description{

}

\usage{
dashJsmole(id=NULL, actionCode=NULL, atomClick=NULL, atomHover=NULL,
atomsMark=NULL, bondClick=NULL, bondHover=NULL,
bondsMark=NULL, height=NULL, jmeEvent=NULL, lineWidth=NULL,
options=NULL, smiles=NULL, src=NULL, width=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{actionCode}{Numeric. set  action code for menu}

\item{atomClick}{Numeric. The atom clicked of component}

\item{atomHover}{Numeric. The highlighted Atom}

\item{atomsMark}{Character. The atoms marked string atom,color}

\item{bondClick}{Numeric. The bond clicked of component}

\item{bondHover}{Numeric. The highlighted bond}

\item{bondsMark}{Character. The bonds marked string bond,color}

\item{height}{Character. The height of rendered componenet}

\item{jmeEvent}{Named list. jsmeEvent information}

\item{lineWidth}{Numeric. The line width of component}

\item{options}{Character. The options for rendered component}

\item{smiles}{Character. The smiles string of component}

\item{src}{Character. The source url for jsme.nocache.js}

\item{width}{Character. The width of rendered compnent}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

